Load classes from their origin classloader
The previous method bootstrap handler used to call `class.forName(String)`
to perform lookups. This causes classes to be loaded from the system
classloader and not from the one of the caller class.

This change uses `class.forName(String, boolean, ClassLoader)`. The correct
classloader is obtained through the lookup class context.